FILE: creates game-setup page component and moves former contents to new game-setup-form component
