+++ libs/spirit-islander/feature-game-setup/src/lib/game-setup-output/game-setup-output.component.html
